Add flow package for auto-numbering using Excel
Add a new Power Automate flow that auto-numbers new tasks added to a Microsoft Planner plan. This flow has the following limitations:
1. It only numbers tasks created after this flow is connected to a plan, not the ones created before.
2. Task numbering might take up to 5 minutes, depending on your Microsoft 365 package
3. Tasks copied over and manually numbered will still be auto-numbered.

The advantage of this approach (over the other flow) is that tasks created within the time window between 2 successive flow runs will not be given the same numbers.